perf-graph-backend/requirements.txt
# Web Framework & API Handling
fastapi~=0.115.5
uvicorn~=0.32.1
gunicorn==23.0.0
grpcio<=1.67.1
httpx~=0.27.2
jiter~=0.8.2
python-multipart>=0.0.20

# Similarity search related
tiktoken>=0.8.0
bs4>=0.0.2
onnxruntime>=1.21.0
pymilvus==2.5.0

# Langchain Ecosystem
langchain~=0.3.20
langchain-core~=0.3.33
langchain-community~=0.3.16
langchain-openai~=0.2.9
langchain-milvus~=0.1.8
langchain-text-splitters>=0.3.5
langsmith~=0.1.145

# Langgraph
langgraph~=0.2.68

# Utils
numexpr~=2.10.2
numpy>=1.21.6
pandas~=2.2.3
pyarrow>=18.1.0
pydantic~=2.10.1
pydantic-settings~=2.6.1
pyowm~=3.3.0
python-dotenv~=1.0.1
setuptools~=75.6.0
streamlit~=1.40.1
pymongo~=4.11.2
typing_extensions~=4.12.2
cachetools~=5.5.2
pypdf~=5.4.0